Fixed issue where automatic choice was deactivated when updated by changelistener
Added resetBookingHappened to keep track of when resetBooking() ran last time ticket-selector was changed.

Added condition to if in SaloonPage getBestSeat() to make sure it doesn't run if booking has just been reset, when bestSeats would otherwise become -1.

Added else if in saloonPage getBestSeats(), so program doesn't try to iterate -1.